--- a/cypress/site/functional/table-of-content.spec.js
+++ b/cypress/site/functional/table-of-content.spec.js
@@ -2,8 +2,10 @@ describe('table of content', () => {
   beforeEach(() => {
     cy.viewport(1600, 660);
     cy.mockConsentAndVisit('/components/tabs');
+    // Hack for Next 13 - this could be removed AFTER upgrading Cypress to v.12
+    // although this statement avoids the need for "cy.wait(1000);" at the end
+    cy.get('section[id="introduction"]').should('be.visible');
     cy.get('[href="/components/tabs/#introduction"]').as('TOCIntroduction');
-    cy.wait(1000);
   });
 
   it('should highlight the first section on load, Introduction', () => {