--- a/site/components/media-list/media-list.tsx
+++ b/site/components/media-list/media-list.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {StyledComponent} from '@emotion/styled';
 import {Cell, Grid, Block, Visible, BreakpointKeys, styled} from 'newskit';
-import LinkNext from 'next/link';
+import NextLink from 'next/link';
 import {UsageCard, UsageCardProps} from '../usage-card';
 import {BaseCard, BaseCardProps} from '../base-card';
 import {MediaListProps} from './types';
@@ -65,9 +65,9 @@ export const MediaList: React.FC<MediaListProps> = ({
           // eslint-disable-next-line react/no-array-index-key
           <Cell {...cellColumnList} key={`index-${index}`}>
             {cardProps.href && (
-              <LinkNext href={cardProps.href} passHref>
+              <NextLink legacyBehavior href={cardProps.href} passHref>
                 {styledCardComponentWithProps}
-              </LinkNext>
+              </NextLink>
             )}
             {!cardProps.href && styledCardComponentWithProps}
           </Cell>