--- a/site/components/menu-collapsible/menu-collapsible.tsx
+++ b/site/components/menu-collapsible/menu-collapsible.tsx
@@ -13,7 +13,7 @@ import {
   TextBlock,
   toNewsKitIcon,
 } from 'newskit';
-import LinkNext from 'next/link';
+import NextLink from 'next/link';
 import {Launch as OutlinedLaunch} from '@emotion-icons/material-outlined/Launch';
 import {IconExpandLess, IconExpandMore} from '../icons';
 import {MenuCollapsible, MobileNavigationDivider, StyledAnchor} from './styled';
@@ -86,7 +86,7 @@ export const MenuMobileCollapsible = ({
         list.map(({title, id, subNav, page}) => (
           <React.Fragment key={id}>
             {page ? (
-              <LinkNext href={id} passHref>
+              <NextLink legacyBehavior href={id} passHref>
                 <MenuItem
                   href={id}
                   selected={path.includes(id)}
@@ -108,7 +108,7 @@ export const MenuMobileCollapsible = ({
                 >
                   {title}
                 </MenuItem>
-              </LinkNext>
+              </NextLink>
             ) : (
               <>
                 <MenuGroup