--- a/site/pages/components/byline.mdx
+++ b/site/pages/components/byline.mdx
@@ -45,7 +45,6 @@ interface BylineData {
   ariaLabel?: string;
 }
       `}
-      }
     </Code>
   </Prop>
   <Prop name="overrides" type="object">