--- a/site/pages/components/consent-settings-link.mdx
+++ b/site/pages/components/consent-settings-link.mdx
@@ -16,7 +16,7 @@ The consent settings link is a component utilising the <a href='/components/link
 
 ## Prerequisites
 
-To utilise this component you will need to have added the <a href='/components/consent'>Consent</a> component added to the head of the page. You will also need to obtain your site id and privacy manager id from Sourcepoint, these are different from the account id used with the [Consent](components/consent) component.
+To utilise this component you will need to have added the <a href='/components/consent'>Consent</a> component added to the head of the page. You will also need to obtain your site id and privacy manager id from Sourcepoint, these are different from the account id used with the [Consent](/components/consent) component.
 
 ## Usage