--- a/site/pages/components/experimentation-web.mdx
+++ b/site/pages/components/experimentation-web.mdx
@@ -35,7 +35,7 @@ It’s important to load the Optimizely script before the utag script, so that t
     <b>scriptCdn</b>.
     <Prop required name="scriptCdn" type="string" required>
       Optimizely script CDN URL.
-      The URL can be found in <i>Settings -> Implementation</i> for every web project in the Optimizely platform.
+      The URL can be found in <i>Settings {'->'} Implementation</i> for every web project in the Optimizely platform.
       More information on how to retrieve the script can be found{' '}
       <a href="https://documentation.sourcepoint.com/web-implementation/sourcepoint-set-up-and-configuration-v2">
         here