--- a/site/pages/index.tsx
+++ b/site/pages/index.tsx
@@ -156,7 +156,7 @@ export async function getStaticProps() {
       JSON.stringify(releasesOrError),
     );
   } else {
-    releases = [];
+    releases = releasesOrError;
   }
 
   const content = {...heroCardFallbackContent, ...formatSheetData(data)};