… package.json
… site/components/feature-card/__tests__/__snapshots__/feature-card.test.tsx.snap
… site/components/sidebar/__tests__/__snapshots__/sidebar-navigation.test.tsx.snap
… site/components/table-of-contents/__tests__/__snapshots__/table-of-contents.test.tsx.snap
… yarn.lock