PPDSC-2518

**What**

1. Background - upgrading to next.js 13 will lay the groundwork for future performance improvements
2. What did you do - upgraded dependencies, tweaked unit tests and e2e tests to still work, measured before and after performance.
3. What does the reviewers should expect

(P1) Performance test based on generating a Lighthouse report via Chrome Developer extensions

**BEFORE** (using http://ncu-newskit-docs.s3-website-eu-west-1.amazonaws.com/ppdsc-2647-image-width-height-props/)
Performance: 86 Accessibility: 87 Best practices: 83 SEO: 100

**AFTER** (using http://ncu-newskit-docs.s3-website-eu-west-1.amazonaws.com/ppdsc-2518-chore-next13/)
Performance: 87 Accessibility: 87 Best practices: 83 SEO: 100

**CONCLUSION**
From a Lighthouse POV, there is not much difference before and after a minimal upgrade to next 13. For more details, see the 2 Lighthouse PDFs attached to the JIRA.

(P2) Performance test based on running all Cypress tests in accessibility.spec.js

**BEFORE**
107 passes, 2 failures in 1062.11s

**AFTER**
107 passes, 2 failures in 938.94s

**CONCLUSION**
Comparing before and after results suggest that Next 13 is about 12% quicker than next 12

(P3) Performance test using Autocannon for 10s on the docs home page (autocannon http://localhost:8081/)



```
Running 10s test @ http://localhost:8081/
10 connections

**BEFORE**
┌─────────┬────────┬────────┬─────────┬─────────┬────────────┬───────────┬─────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%   │ 99%     │ Avg        │ Stdev     │ Max     │
├─────────┼────────┼────────┼─────────┼─────────┼────────────┼───────────┼─────────┤
│ Latency │ 905 ms │ 952 ms │ 2012 ms │ 2021 ms │ 1086.25 ms │ 333.06 ms │ 2021 ms │
└─────────┴────────┴────────┴─────────┴─────────┴────────────┴───────────┴─────────┘
┌───────────┬─────┬──────┬─────────┬─────────┬─────────┬────────┬─────────┐
│ Stat      │ 1%  │ 2.5% │ 50%     │ 97.5%   │ Avg     │ Stdev  │ Min     │
├───────────┼─────┼──────┼─────────┼─────────┼─────────┼────────┼─────────┤
│ Req/Sec   │ 0   │ 0    │ 10      │ 10      │ 9       │ 3      │ 10      │
├───────────┼─────┼──────┼─────────┼─────────┼─────────┼────────┼─────────┤
│ Bytes/Sec │ 0 B │ 0 B  │ 2.75 MB │ 2.75 MB │ 2.48 MB │ 826 kB │ 2.75 MB │
└───────────┴─────┴──────┴─────────┴─────────┴─────────┴────────┴─────────┘

Req/Bytes counts sampled once per second.
# of samples: 10

100 requests in 10.02s, 24.8 MB read
```

**AFTER**
```
Running 10s test @ http://localhost:8081/
10 connections


┌─────────┬────────┬────────┬─────────┬─────────┬───────────┬───────────┬─────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%   │ 99%     │ Avg       │ Stdev     │ Max     │
├─────────┼────────┼────────┼─────────┼─────────┼───────────┼───────────┼─────────┤
│ Latency │ 732 ms │ 805 ms │ 1230 ms │ 1232 ms │ 849.52 ms │ 128.06 ms │ 1233 ms │
└─────────┴────────┴────────┴─────────┴─────────┴───────────┴───────────┴─────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬────────┬────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg    │ Stdev  │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼────────┼────────┼─────────┤
│ Req/Sec   │ 10      │ 10      │ 10      │ 16      │ 11     │ 2.05   │ 10      │
├───────────┼─────────┼─────────┼─────────┼─────────┼────────┼────────┼─────────┤
│ Bytes/Sec │ 2.75 MB │ 2.75 MB │ 2.95 MB │ 4.71 MB │ 3.2 MB │ 627 kB │ 2.75 MB │
└───────────┴─────────┴─────────┴─────────┴─────────┴────────┴────────┴─────────┘

Req/Bytes counts sampled once per second.
# of samples: 10

120 requests in 10.02s, 32 MB read
```

**CONCLUSION**
Comparing before and after results suggest that Next 13 is about 20% quicker than next 12

<!---
This section will be used to indicate if we should move to a major version in the next release, remove if not revelant.
DO CONSIDER any of the following are breaking changes to a consumer, this is by no mean an exhaustive list.
-removing or renaming props
-removing or renaming tokens
-removing or renaming components
-removing or renaming exported functions
-in some cases, major bumps to peer dependencies
--->

<!---
Add any breaking change if present.
E.g:
BREAKING CHANGE: renames the foobar component's prop foo to bar
--->

**I have done:**
- [x] Written unit tests against changes
- [x] Written functional tests against the component and/or NewsKit site
- [ ] Updated relevant documentation

**I have tested manually:**
- [x] The feature's functionality is working as expected on Chrome, Firefox, Safari and Edge
- [ ] The screen reader reads and flows through the elements as expected.
- [x] There are no new errors in the browser console coming from this PR.
- [ ] When visual test is not added, it renders correctly on different browsers and mobile viewports (Safari, Firefox, small mobile viewport, tablet)
- [x] The Playground feature is working as expected


<!---
Below sections are optional
--->

**Before:**
<!--- Drag and Drop your screenshot's here --->

**After:**
<!--- Drag and Drop your screenshot's here --->

**Who should review this PR:**
<!---
If you know someone is a domain expert for your PR,
someone who is deeply involved in the story,
ask them explicitly to review the PR.
--->

**How to test:**
<!--
If it's not immediately obvious how to test this PR, give instructions.
It's mandatory to update README.MD or development documentation if existing test strategy had changed.
-->

<!--
More info about raising an good PR: https://nidigitalsolutions.jira.com/wiki/spaces/NPP/pages/1319370846/Pull+Request
-->